⭐⭐⭐ C357: the branch's entire firmware warning debt was two lines — now fixed
C356 could only reach the 14 host-compilable files and said so. lf_reader_main.c,
lf_indala_data.c and app_cmd.c need the Nordic headers, and they carry every
writer and the whole capture engine. This measures them, by building the real
firmware with -Wconversion added after -Werror.

1895 conversion warnings firmware-wide, overwhelmingly the SDK. Inside src/rfid
the HF stack dominates and none of it is ours — mf1_toolbox.c 31, nfc_mf1.c 18,
nfc_mf0_ntag.c 17, rc522.c 16, nfc_seos.c 10, all pre-existing.

Not one of this branch's new LF files appears at all: lf_indala_psk.c,
lf_slicer.c, lf_ask_manchester.c, lf_ask_biphase.c, lf_fsk2a.c, and
lf_indala_data.c — the very file C356 could not reach.

lf_reader_main.c had three, and git blame splits them:

  line 337   098e0a91   on main — upstream's
  line 359   aff5e377   ours, from C326 (the password fix)
  line 370   aff5e377   ours, from C326

All three narrow bytes_to_num(ptr, 4) to uint32_t, which is provably safe at
len 4. Our two now carry an explicit cast saying so, verified 3 → 1 with the
survivor being upstream's. Upstream's is left alone deliberately: fixing it would
widen F1's diff with a change that is not F1, and F1 is meant to be the first PR
a maintainer sees.

A file-level count would have charged us all three. Blame is what separates ours
from inherited, and it is the difference between "the branch adds warnings" and
"the branch added two lines and fixed them".

Also recorded so a PR does not propose it: -Wextra is not usable on this
codebase. It produces 180 "unused parameter 'conn_handle'", 132 'status', 103
'length', 77 'data' — the command-dispatch signature every handler must take and
most do not fully use. That is an API shape, not a defect, and it is why the
project builds at -Wall.

Makefile restored, make check green, firmware builds.

+++ firmware/application/src/rfid/reader/lf/lf_reader_main.c
@@ -356,7 +356,10 @@ static void try_reset_t55xx_passwd(uint32_t new_passwd, uint8_t *old_passwds, ui
  *   2. the new password, only when one was actually supplied;
  *   3. an open write LAST, so an unprotected tag finishes correctly. */
 static uint8_t write_t55xx(uint32_t *blks, uint8_t blk_count, uint8_t *new_passwd, uint8_t *old_passwds, uint8_t old_passwd_count) {
-    const uint32_t passwd = bytes_to_num(new_passwd, 4);
+    /* ⚠ `bytes_to_num` returns uint64_t; at len 4 the value cannot exceed 32 bits, so the
+     * narrowing is provably safe. Cast explicitly to say so — it is the only -Wconversion
+     * warning this branch added to the firmware (C357). */
+    const uint32_t passwd = (uint32_t)bytes_to_num(new_passwd, 4);
     const bool set_pwd = (passwd != 0);
 
     if (set_pwd && blk_count > 0) {
@@ -367,7 +370,7 @@ static uint8_t write_t55xx(uint32_t *blks, uint8_t blk_count, uint8_t *new_passw
     bsp_delay_ms(1);  // Delays for a while after starting the field
 
     for (uint8_t k = 0; k < old_passwd_count; k++) {
-        const uint32_t old = bytes_to_num(old_passwds + k * 4, 4);
+        const uint32_t old = (uint32_t)bytes_to_num(old_passwds + k * 4, 4);
         if (old == 0) {
             continue;
         }
